Removed the functionality for setting log levels; Reordered the test project;
