+++ PostSharpTutorial/LoggerAspect.Tests/LoggerAspect.Tests.csproj
@@ -73,6 +73,7 @@
     <Compile Include="LoggingAspectBaseFunctionalityTests.cs" />
     <Compile Include="MockLogger.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TestClasses.cs" />
     <Compile Include="TestExcludeProperties.cs" />
   </ItemGroup>
   <ItemGroup>
