+++ PostSharpTutorial/LoggerAspect.Tests/TestExcludeProperties.cs
@@ -18,8 +18,9 @@ namespace LoggerAspect.Tests
         }
 
         [Test]
-        public void WhenAppliedToClassWithNoExclude_ShouldLogProperty()
+        public void WhenAppliedToClassWithNoExclude_ShouldLogPropertyAndConstructor()
         {
+
             // act
             var p = new Person {Name = Guid.NewGuid().ToString()};
 
@@ -112,119 +113,5 @@ namespace LoggerAspect.Tests
                 .Be(0, "because we do not hit the Debug method for constructors and properties");
         }
 
-    }
-
-    [LoggingAspect]
-    public class Person
-    {
-        public string Name { get; set; }
-    }
-
-    [LoggingAspect(Exclude = ExclusionFlags.Properties)]
-    public class PersonExcludeProperty
-    {
-        public string Name { get; set; }
-    }
-
-    [LoggingAspect(Exclude = ExclusionFlags.InstanceConstructors)]
-    public class PersonExcludeInstanceConstructor
-    {
-        public PersonExcludeInstanceConstructor()
-        {
-        }
-
-        public string Name { get; set; }
-    }
-
-    [LoggingAspect(Exclude = ExclusionFlags.StaticConstructor)]
-    public class PersonExcludeStaticConstructor
-    {
-        static PersonExcludeStaticConstructor()
-        {
-
-        }
-
-
-        public string Name { get; set; }
-    }
-
-    [LoggingAspect(Exclude = ExclusionFlags.PropertySetters)]
-    public class PersonExcludePropertySetters
-    {
-        static PersonExcludePropertySetters()
-        {
-
-        }
-
-        public PersonExcludePropertySetters()
-        {
-        }
-
-        public PersonExcludePropertySetters(string name)
-        {
-            Name = name;
-        }
-
-        public string Name { get; set; }
-    }
-
-    [LoggingAspect(Exclude = ExclusionFlags.PropertyGetters)]
-    public class PersonExcludePropertyGetters
-    {
-        static PersonExcludePropertyGetters()
-        {
-
-        }
-
-        public PersonExcludePropertyGetters()
-        {
-        }
-
-        public PersonExcludePropertyGetters(string name)
-        {
-            Name = name;
-        }
-
-        public string Name { get; set; }
-    }
-
-    [LoggingAspect(Exclude = ExclusionFlags.Constructors)]
-    public class PersonExcludeConstructors
-    {
-        static PersonExcludeConstructors()
-        {
-
-        }
-
-        public PersonExcludeConstructors()
-        {
-        }
-
-        public PersonExcludeConstructors(string name)
-        {
-            Name = name;
-        }
-
-        public string Name { get; set; }
-    }
-
-    [LoggingAspect(Exclude = ExclusionFlags.Properties | ExclusionFlags.Constructors)]
-    public class PersonExcludePropertyConstructors
-    {
-        static PersonExcludePropertyConstructors()
-        {
-
-        }
-
-        public PersonExcludePropertyConstructors()
-        {
-        }
-
-        public PersonExcludePropertyConstructors(string name)
-        {
-            Name = name;
-        }
-
-        public string Name { get; set; }
     }
 }
